fixing jacoco tests. Refactoring Project structure
