+++ GameOfLife/android/src/com/mygdx/gameoflife/AndroidLauncher.java
@@ -1,10 +1,8 @@
 package com.mygdx.gameoflife;
 
 import android.os.Bundle;
-
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
-import com.mygdx.gameoflife.src.GameOfLife;
 
 public class AndroidLauncher extends AndroidApplication {
 	@Override
