+++ GameOfLife/core/src/com/mygdx/gameoflife/main/com/mygdx/gameoflife/core/Player.java
@@ -1,4 +1,4 @@
-package com.mygdx.gameoflife.src.core;
+package com.mygdx.gameoflife.core;
 
 import com.badlogic.gdx.graphics.Color;
