+++ GameOfLife/core/src/com/mygdx/gameoflife/main/com/mygdx/gameoflife/screens/JoinGameScreen.java
@@ -1,4 +1,4 @@
-package com.mygdx.gameoflife.src.screens;
+package com.mygdx.gameoflife.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
@@ -25,8 +25,8 @@ import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.StretchViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
-import com.mygdx.gameoflife.src.GameOfLife;
-import com.mygdx.gameoflife.src.core.AvailableServerDetails;
+import com.mygdx.gameoflife.core.AvailableServerDetails;
+import com.mygdx.gameoflife.GameOfLife;
 
 import java.util.ArrayList;
 import java.util.List;
