+++ GameOfLife/core/src/com/mygdx/gameoflife/main/com/mygdx/gameoflife/screens/MainMenuScreen.java
@@ -17,9 +17,8 @@ import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.StretchViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
+import com.mygdx.gameoflife.core.Player;
 import com.mygdx.gameoflife.GameOfLife;
-import com.mygdx.gameoflife.src.core.Player;
-import com.mygdx.gameoflife.src.screens.StartGameScreen;
 
 import java.util.ArrayList;
