+++ GameOfLife/core/src/com/mygdx/gameoflife/main/com/mygdx/gameoflife/screens/StartGameScreen.java
@@ -1,4 +1,4 @@
-package com.mygdx.gameoflife.src.screens;
+package com.mygdx.gameoflife.screens;
 
 
 import com.badlogic.gdx.Gdx;
@@ -22,8 +22,8 @@ import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.StretchViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
-import com.mygdx.gameoflife.src.GameOfLife;
-import com.mygdx.gameoflife.src.core.Player;
+import com.mygdx.gameoflife.core.Player;
+import com.mygdx.gameoflife.GameOfLife;
 
 public class StartGameScreen implements Screen {
     private OrthographicCamera gameCamera;
