+++ GameOfLife/core/src/com/mygdx/gameoflife/test/com/mygdx/gameoflife/core/AvailableServerDetailsTest.java
@@ -1,9 +1,9 @@
-package com.mygdx.gameoflife.test.core;
+package com.mygdx.gameoflife.core;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import com.mygdx.gameoflife.src.core.AvailableServerDetails;
+import com.mygdx.gameoflife.core.AvailableServerDetails;
 
 import org.junit.Test;
