+++ GameOfLife/core/src/com/mygdx/gameoflife/test/com/mygdx/gameoflife/mock/TestBoard.java
@@ -1,6 +1,6 @@
-package com.mygdx.gameoflife.test.mock;
+package com.mygdx.gameoflife.mock;
 
-import com.mygdx.gameoflife.src.core.Board;
+import com.mygdx.gameoflife.core.Board;
 
 public class TestBoard extends Board {
